feat(cli): initial draft for CLI
This is the initial draft for the CLI. Implemented functionality to
read a CSV File as Dict, create a list of People and Errors, and
write these objects out to respective files. The input and output
files are hardcoded.

Added base level documentation in the form of a README.md, along
with appropriate levels of documentation for each method and class.